Implement POC to extract a Ladder diagram from 'Decide' statements
A Ladder diagram is a visualisation of a boolean circuit, which we can
visualise with various javascript libraries.

We provide a very simple implementation, where a client can request a
ladder diagram for all 'Decide' statements in a file. We produce JSON
which can be taken directly to visualise the boolean circuit
representing our program.