Extract the core of HLS into a package
This commit copies parts of the HLS codebase, to be precise ghcide, into
the mattwaddington codebase.
However, since jl4 is not a Haskell derivate, we stripped everything out
of the codebase that is ghcide specific.

This includes (and likely not limited to)

* LSP Config specific to ghcide
* HieDB, which is deeply wired into ghcide at this point
* SessionLoading, which is the `hie-bios` component responsible for
  initialising a GHC Session
* Rules for implementing a GHC frontend
  * Such as Typechecking, desugaring, template-haskell, etc...

Additionally, removed the `opentelemetry` support. Even though, it is
not specific to GHC and might be helpful in practice, it does complicate
the codebase. Basically, it was removed because fendor does not believe
the opentelemetry support carries its own weight.

We also moved the `Ide.Logger` module into this sub-package, as we do
need the Logger in this sub-package, without depending on
`hls-plugin-api`.

As a notable implementation change, we had to rip out mentions of ghcide
specific `Config`. While we could have parameterised the types by
`config`, this felt like an intrusive refactor, complicating major parts
of the API.
Thus, we replace `LanguageContextEnv Config` with a specially crafted
record 'LspSink' which contains monomorphed versions of LSP specific functions
that are used throughout the LSP.Core package.