Add Language Server for jl4
Re-implement the Language Server for JL4 language based on the LSP
implementation based off the ghcide codebase.

Set up the Language Server similarly to ghcide, but make some
simlifications, in particular, we do not have a plugin system, but write
in a monolithic style.

Implements rules for lexing, parsing and typechecking.