Implement SemanticToken highlighting with fallback
Syntax highlighting has two stages of colouring:

* Based on lexed tokens
* Based on the parsed AST

When parsing fails, there was previously no syntax highlighting at all,
which looks awkward in and incorrect.
To improve the situation we try to insert Semantic Tokens using only the
tokens available after lexing.
This means we are losing a information, but we have the advantage that
we still provide syntax highlighting when the file doesn't parse at this
exact moment.

Especially helpful if the file doesn't parse and the user opens the file
in the editor for the first time.