Implement 'CodeAction' to 'ASSUME' unknown vars as given
When a user writes an unknown variable, offer a 'CodeAction' which
adds an 'ASSUME' statement above the top-decl with the inferred type.

For the implementation, we are adding the structured diagnostic to
'FileDiagnostic' so that we can later use the structured error message
to produce the 'CodeAction'.
We store the structured message as a 'NfDynamic' which is the same as a
'Dynamic' with an additional 'NFData' constraint.
We add a couple of convenience pattern matching functions to not force
users to interact with the 'NfDynamic' directly.