Add source spans to 'Anno_' and 'AnnoElement_'
Source spans (also called source ranges or 'SrcRange') are an integral
part of our framework.
We often need to know the location for error reporting, but also for
finding context information.
Moreover, they are going to be important for implementing refactorings,
such as turning unknown variable errors into Code Actions 'ASSUME' that
a variable is actually known.

In this commit, we enhance 'Anno_' and 'AnnoElement_' with the cached
source ranges of their AST node. In particular, 'Anno_' caches the
source Range of all of its child 'AnnoElement_' and each 'AnnoElement_'
caches the range of its 'CsnCluster'. For 'AnnoHole_', we expect to
receive a source range hint. This can be usually obtained from the child
AST node's 'Anno_' element, which has its source range cached already.

For the future: 'setAnno' can be a dangerous function, as the caller
needs to remember to recompute the source range of the 'Anno_'.