Attach 'Nlg' annotations in-tree of 'Program Name'
We want to write Nlg (natural language generation) annotations directly
in our program which are then smartly attached to the best fitting syntax
node. At the moment, we only support Haddock style Nlg comments, where the
comments must be in a trailing position.

For example:

```
DECIDE foo [Better Name] IS 5
```

We want the Nlg annotation '[Better Name]' to be attached to 'foo'.
This annotation can be used for hover information (e.g., whenever we hover over any
use-sites of 'foo'), but is also supposed to be used when we linearise
our program to human readable text.

Virtually all 'Name's can have such a NLG annotation, including Names
found in `DECLARE`, `GIVEN`, `GIVETH`, `DECIDE`, `ASSUME`, `CHECK`,
etc...
Even in sub-expressions can individual uses of 'Name's have a Nlg
Annotation.

The process of attaching Nlg annotations directly into our AST consists
of three major steps:

1. Add 'Nlg' type and add it to the 'Anno.extra'

We introduce a new 'Extension' type which carries both 'Nlg' and 'Type' information.
Eventually, we want to have 'Nlg' annotations for virtually any
sub-term, so it makes sense to store the 'Nlg' in the 'Anno'.

2. Adapt the parser to save 'Nlg' comments for later processing

We change our 'Parser' type to store information about 'Nlg's and 'Ref's
in a separate state, called 'PState'. The base 'Parser' type is turned
into a monad transformer over a 'StateT'.
Each 'Nlg' (or 'Ref') is its own 'CsnCluster', thus, it also stores
trailing whitespace tokens and comments.
To maintain exactprinting capabilities, we still add the
parsed 'Nlg's as 'CsnCluster's to the 'Anno' of the respective syntax
node we are currently parsing.
However, we mark them as 'Hidden', to indicate they are not part of the
abstract syntax tree and should be ignored for operations that operating
on the abstract syntax, not the concrete syntax.
In particular, we ignore the hidden 'CsnCluster's in 'HasSrcRange', to still
provide accurate 'SrcRange's for 'Name's, etc...

3. Attach 'Nlg' annotation to the correct 'Name'

Since we are stealing the Haddock Syntax, we generously borrow GHC's
implementation of attaching Haddock comments to the ast.
We implement an ast traversal which keeps track of the 'SrcRange' of
the currently processed ast node. By utilising the 'Applicative'
instance, this allows us to relatively trivially find the 'SrcRange' of
potential Nlg annotations for any 'Name' in the ast.

The implementation itself is rather tricky, thus we copied a lot of the
documentation from `compiler/GHC/Parser/PostProcess/Haddock.hs` [1], but we
might be missing some details.
When in doubt, take a look at the GHC implementation, which might give
additional justification for certain design decisions.

---

At last, we report the warnings that have been produced while attaching the
annotations onto our 'Program Name'.

[1] https://gitlab.haskell.org/ghc/ghc/-/blob/b1a2c5e4508d61bc0bedc13c7240b6fbf674157e/compiler/GHC/Parser/PostProcess/Haddock.hs