Exclude hidden concrete syntax nodes from Anno SrcRange calculation
When deciding the `SrcRange` of an `Anno`, we are already doing some
semi arbitrary filtering of the things we count towards the overall
`SrcRange`.
For example, we exclude the `SrcRange` of any trailing tokens.
Now, we also exclude any `CsnCluster` which isn't visible in the AST,
but only part of the CST (concrete syntax tree).

This gives us the "proper" `SrcRange` of what we consider to be part of
the individual AST node.