Introduce naive linearization of JL4 expressions
Linearization is the process of translating an abstract syntax tree into
a readable version of it.

Our first approximation, introduces the 'LinTree' type, which is a flat
representation of tokens produced from linearizing an 'Expr Resolved'.
We do not attempt to produce truly human-readable text, but rather try
to be predictable.