Split JL4 specific and lsp generic setup code
Since we would like to use the lsp code for other projects as well, we
need to be able to re-use as much as possible of the language server.
This commit splits up jl4 specific code and generic lsp setup code.

The intention is to re-use the generic lsp setup code for other
languages.