[feat] lookup references
New rule which creates a mapping from source locations to the defining
occurence of a name and a mapping from defining occurences to source
ranges of references of that occurence.

The handler builds the mapping based on the rule and looks up the
definition in the map by the use of the source position sent as part of
the lookup.