Make equality polymorphic.
This replaces equality with a datatype-generic implementation
which fails at run-time for unsupported types (i.e., functions).

No change yet for comparison operators, although we could
potentially (with more work) do something similar for these
as well.

Also fixes a small bug in the unification / occurs check that
I had not noticed earlier.

Also adds a test case implementing a membership test on lists
and testing it on various different lists. This was tricky before
because one would have to pick a concrete list element type.