Add more hover info, and improve type error messages.
This does quite a few things:

1. Add the possibility to have kind info on hover, and implement this.
2. Refactor the imports of various modules.
3. Change the type error messages to have significantly more detail.
4. Remove simpleprint and replace it with fendor's prettyLayout.
5. Fix range calculation for hover info.
6. Add hover info for term-level ASSUMEs.

Additional info about 5:

This calculation predated ranges directly stored in annotations,
therefore it tried to compute the range itself in an ad-hoc fashion.

But the way it did it, it would just merge the ranges of the
non-anno children, which isn't correct if the anno contains the
first token (i.e., a keyword).

Fixes #165.

Thanks to Magnus for reviewing.