Extend the parser to further process NLG annotations
So far, NLG annotations were treated as single tokens that we do not
look at in detail. This is insufficient, as the user can provide a
custom NLG annotation which references variables, delimited by `%`,
that are then incorporated by the NLG linearization step.

For example:

    @nlg function %a% takes parameter %`the param`%

This NLG annotation references the variable `a` and the variable
`` `the param` ``.
Since we are already parsing the names, we can also provide better
syntax highlighting for the `Name`s that occur in an NLG annotation.

Thus, we perform ad-hoc lexing and parsing of NLG annotation during the parser
step, where we replace the single NLG token with the parsed result.
The new NLG annotations store which `Name`s they are referencing,
s.t., we can later perform scope checking of the names references in the
NLG annotation.

In the typechecker, we need to perform scope checking at the right
locations to make sure all variables are actually in scope.
This is currently rather hackily done, and likely needs some more
thoughts.