Add decision web-service which allows you to run L4 functions through a REST API (#205)
* Factor out SrcRange and SrcPos to break cyclic module imports

* Add example programs for the webservice

* Add decision web-service which allows you to run L4 functions through a REST API

This ports the web-service from the smucclaw/dsl repository.
As such, it contains a couple of design decision which made sense for
the dsl repo, but do not make a lot of sense right now.

For now, I deliberately decided to keep old code around, for the
non-zero chance that we might need this code again.

In short, this adds a REST service where you can add scenarios/functions
which can later be evaluated using REST endpoints.
As Jl4 does not support some of the features simala or gml do support,
we have some arbitrary restrictions:

* No floating point numbers
* No objects
* No uncertain
* No unknown

There will also be no reasoning tree, as Jl4 doesn't have the tracing
feature of simala, nor the explanation part of gml.

Additionally, the Jl4 endpoint can not handle missing values, and will
always throw an error when trying to invoke a function without all
required values.

We additionally ship a swagger-ui which is served under:

* GET /swagger-ui
* GET /swagger.json

* Add lzma to GitHub container

* [feat] deploy on /decision and apply fixes

---------

Co-authored-by: mangoiv <contact@mangoiv.com>