Expose NLG information in Hover text
We expose the NLG linearisation to the `InfoTree`. This makes it easy to
discover and can be used as a form of documentation.

Additionally, add a bunch of type and name resolution to the AST after
type checking.

Similarly to the NLG annotations, we need to be careful to add the
additional type information when the respective type information is
actually available.
For example, when new variables are brought into scope, for example in:

```
  ...
  CONSIDER x
  WHEN y FOLLOWED BY ys THEN y
  OTHERWISE 0
```

then the `Resolved` `y` is only in scope within the `inferBranch` code,
i.e. `THEN y` part. Thus, we need to make sure to add type information
to `y` after we have typechecked the body of branch.

The same logic is applied to anything that brings a new binder into
scope.
As a rule of thumb, whenever we have a `scope` within the typechecker,
some extra logic will need to be added.