Refactor TypeCheck phase to use explicit Scopes
Instead of relying on mutations in the 'TypeCheck' phase, we rather want
to explicitly keep track of which Identifiers are in actually in scope.

From a technical point of view, we don't keep track of 'Environment' and
'CheckEntity' in a State monad, which requires us to explicitly use
scoping functions, such as 'scope' and keep track of exposed
Identifiers, such as function names.

This is less elegant and makes the scopes much harder track. Instead, we
keep the 'Environment' and 'CheckEntity' in a 'Reader'.
The Reader has a natural scoping mechanism 'local' and delimits the
scope of Identifiers more naturally.

We still need to explicitly keep track of Identifiers that need to be
exported to the outside scope, but we do so now in a consistent way,
instead of shotgun changes.

Most of the changes are rather mechanical, and not exciting.