Implement Semantic Syntax Highlighting based on TypeCheck results
Typechecker results can add additional semantic tokens to the syntax
highlighting.

We add a new 'Rule' `TypeCheckedSemanticTokens` which generates
'SemanticTokens' based on 'Module Resolved'.
These tokens are merged with the result of `ParserSemanticTokens` and
`LexerSemanticTokens`, to provide the most accurate and helpful syntax
highlighting.

We have to very careful when modifying or generating AST nodes during the typecheck
phase to not accidentally reuse `Anno`s of `Name`s or similar.

For example, given this L4 code:

```
GIVEN a
foo MEANS a
```

The parsed AppForm is 'MkAppForm ... "foo" []'. But during typechecking,
we normalise this, and add the "a" from the "TypeSig" to the "AppForm",
resulting in an AST node like: 'MkAppForm ... "foo" ["a"]'.
In this situation, we need to make sure we remove the 'Anno' associated
with the "a", otherwise, during syntax highlighting, we will write out
the "a" after "foo" in the semantic tokens, leading to incorrect syntax
highlighting.