Implement floating point literals
We change the 'NUMBER' type to work with rational numbers instead of
only integers.
This allows us to use floating point literals, such as `2.0`, and `2.3`,
and we still support integer literals, e.g. `1`, `2`.

Since 'NUMBER' is now backed by `Rational`, we cannot represent
irrational numbers accurately, but at least to some approximation.
The advantage of `Rational` over `Double` is that we do not suffer from
IEEE decisions, such as `1.0 / 0.0 = NaN` (or `NaN` in general) and the
usual floating point inaccuracies such as `0.1 + 0.2 = 0.3` is indeed
`TRUE` in L4.

We introduce the following functions for working with 'NUMBER'

* '\`IS INTEGER\`': returns 'TRUE' iff the number is an integer
* 'ROUND': rounds to the closest whole integer
* 'FLOOR': truncates to the next smallest integer
* 'CEILING': increases to the next largest integer

These functions are all built-in into the typechecker and evaluator.

Many other useful utility can be added in user-land or added to the
evaluator on a by-need basis.

Implementation wise, most of the patch is rather straight forward, we
introduce a new TokenType 'TFloatLit' which parses numbers such as '2.0'
and '1.23456789'.
All 'NUMBER's are now 'Rational', so 'TIntLit' is also converted to
'Rational' during parsing, which is a lossless operation.

In the typechecker, we introduce the new built-in functions, which is
mostly boilerplate.

Similarly in the evaluator, we provide the implementations for the
built-in functions.

Bug fixes included in this patch that are not 100% related to the
feature itself:

* Division by zero throws an exception instead of halting the language
  server
* Module by zero throws an exception instead of halting the langauge
  server