Introduce module for AnnoParser
The pattern 'Compose Parser WithAnno' is used a lot.
Thus, we make it a proper module!
The `AnnoParser` is a Parser (or anything that's a functor) that allows
us to track and compose source annotations.

Also, we introduce versions for 'sepBy' and 'sepBy1' that are type safe
and cannot be misused.